RavenDB-26387 Split components
